requirements.txt
pandas==2.2.1
matplotlib
numpy
mysql-connector-python